refactor:[리팩토링] Course 디렉토리 추가 및 Course 관련 Entity 추가
