+++ footprint/src/main/java/com/umc/footprint/src/walks/WalkController.java
@@ -98,38 +98,4 @@ public class WalkController {
         }
     }
 
-    @ResponseBody
-    @PostMapping("/list")
-    public BaseResponse<List<GetCourseListRes>> getCourseList(@RequestBody String request) throws BaseException, JsonProcessingException {
-
-            // userId(구글이나 카카오에서 보낸 ID) 추출 (복호화)
-            String userId = jwtService.getUserId();
-            log.debug("유저 id: {}", userId);
-            // userId로 userIdx 추출
-            int userIdx = userProvider.getUserIdx(userId);
-
-            GetCourseListReq getWalkListReq = new ObjectMapper().readValue(request, GetCourseListReq.class);
-
-            List<GetCourseListRes> courseList = walkService.getCourseList(getWalkListReq,userIdx);
-
-            return new BaseResponse<>(courseList);
-    }
-
-
-
-
-    @ResponseBody
-    @GetMapping("/{courseIdx}/infos")
-    public BaseResponse<GetCourseInfoRes> getCourseInfo(@PathVariable int courseIdx){
-
-        try {
-            GetCourseInfoRes courseInfo = walkService.getCourseInfo(courseIdx);
-
-            return new BaseResponse<>(courseInfo);
-        } catch(BaseException exception) {
-            return new BaseResponse<>((exception.getStatus()));
-        }
-
-    }
-
 }
